Fixed compatibility with earlier projects
Instead of modifying the rotation field to float, added a new float
field rotationF - this allows Java's object stream to load the project
files without throwing a tantrum.
The code later updates the rotationF field in loaded FTLGib and FTLMount
instances with the old int value, should the rotationF field be 0.